requirements.txt
django==2.2.2
pytz==2019.1
sqlparse==0.3.0
django-taggit==1.1.0
psycopg2=2.8.3
pillow=6.1.0
